Rename script name
