+++ docs/training/embeddings.md
@@ -17,7 +17,7 @@ torch.save('enc_embeddings.t7', embeddings)
 
 where `embeddings[i]` is the embedding of the \(i\)-th word in the vocabulary.
 
-To automate this process, OpenNMT provides a script `tools/embedding_convert.lua` than can download pretrained embeddings from [Polyglot](https://pypi.python.org/pypi/polyglot) or convert trained embeddings from word2vec or GloVe with regard to the word vocabularies generated by `preprocess.lua`.
+To automate this process, OpenNMT provides a script `tools/embeddings.lua` than can download pretrained embeddings from [Polyglot](https://pypi.python.org/pypi/polyglot) or convert trained embeddings from word2vec or GloVe with regard to the word vocabularies generated by `preprocess.lua`.
 
 !!! note "Note"
     The script requires the `lua-zlib` package.
@@ -25,7 +25,7 @@ To automate this process, OpenNMT provides a script `tools/embedding_convert.lua
 For example, to generate pretrained English words embeddings:
 
 ```bash
-th tools/embedding_convert.lua -auto_load en -dict_file data/demo.src.dict -save_data data/demo-src-emb
+th tools/embeddings.lua -auto_load en -dict_file data/demo.src.dict -save_data data/demo-src-emb
 ```
 
 !!! note "Note"
@@ -34,7 +34,7 @@ th tools/embedding_convert.lua -auto_load en -dict_file data/demo.src.dict -save
 Or to map pretrained word2vec vectors to the build vecobulary
 
 ```
-th tools/embedding_convert.lua -embed_type word2vec -embed_file data/GoogleNews-vectors-negative300.bin -dict_file data/demo.src.dict -save_data data/demo-src-emb
+th tools/embeddings.lua -embed_type word2vec -embed_file data/GoogleNews-vectors-negative300.bin -dict_file data/demo.src.dict -save_data data/demo-src-emb
 ```
 
 ## Fixed
